virtualdeb: Add multiarch library paths if present
If usr/lib/$(uname -m)-linux-gnu exists, add that as a preferred direcory.

FIXME: This is incomplete, and might not work on some less popular systems.

The full spec is at https://wiki.debian.org/Multiarch/Tuples -- we rely on
$(uname -m) to resolve to the architecture part of this tuple, but it is
not true on all architectures (though it works for x86_64); and we hardcode
the -linux-gnu suffix.